questlog/back/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
sqlparse==0.4.4
tzdata==2024.1
gunicorn
